requirements.txt
Pillow
openai
mss